# backend/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Swagger demo for the item REST api TP.",
      "version": "1.0.0",
      "title": "Swagger Event API",
      "contact1": {
        "email": "8170308@estg.ipp.pt"
      },
      "contact2": {
        "email": "8170110@estg.ipp.pt"
      },
      "contact3": {
        "email": "8180003@estg.ipp.pt"
      }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": ["http"],
    "paths": {
...
      "/locations/": {
        "get": {
          "tags": ["Location"],
          "summary": "Gets all locations",
          "description": "Gets all locations in the database",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "location List",
              "schema": {
                "$ref": "#/definitions/Location"
              }
            }
          },
          "security": [
            {
              "item_jwt_auth": []
            }
          ]
        }
      },
...
      "/auth/register-promoter": {
        "post": {
          "tags": ["Promoter"],
          "summary": "Create promoter",
          "description": "Creates a promoter only a Admin can do",
          "operationId": "createPromoter",
          "produces": ["application/xml", "application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Created user object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User",
                "role": "PROMOTER"
              }
            }
          ],
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "item_jwt_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    },
    "definitions": {
...
      "Location": {
        "type": "object",
        "required": ["place", "max", "limit"],
        "properties": {
          "_id": {
            "type": "string"
          },
          "place": {
            "type": "string",
            "example": "Rua do Curral"
          },
          "max": {
            "type": "number",
            "example": 10000
          },
          "limit": {
            "type": "number",
            "example": "50",
            "description": "% between 0 and 100"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": ["email", "password", "name", "role"],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "user"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["USER", "PROMOTER", "ADMIN"]
          }
        }
      }
    }
  }